Added User control with a basic Database.
Created DB manager file.
Created DB file.
Created DB Schema.
Added Session control to main webserver.
Added login, logout and register functionalities.
Added list vms by user and list all vms functionalities.

+++ code/webserver/server.py
@@ -1,27 +1,133 @@
-from flask import Flask, render_template, redirect, url_for, request, flash
+from flask import Flask, render_template, redirect, url_for, request, flash, session
 from resourcemanager.mainfunctions import VirtInstance
+import db
 import os
 from time import sleep
+from werkzeug.security import check_password_hash, generate_password_hash
+
 
 app = Flask(__name__)
 app.secret_key = os.urandom(16)
+app.config["DATABASE"] = os.path.dirname(os.path.abspath(__file__))+"/flask.sql"
 libvirt_instance = VirtInstance()
+
+if not os.path.isfile(app.config["DATABASE"]):
+    db.init_db(app)
+    with app.app_context():
+        myDB = db.get_db()
+        myDB.execute('INSERT INTO user (username, password) VALUES (?, ?)',
+                     ("admin", generate_password_hash("admin")))
+        myDB.commit()
+
+db.init_app(app)
 
 
 @app.route("/")
 def hello_world():
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
     return redirect(url_for("list_vms"))
-    # return render_template('index.html')
+
+
+@app.route("/login", methods=('GET', 'POST'))
+def login():
+    if "user_id" in session:
+        return redirect(url_for('hello_world'))
+
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        myDB = db.get_db()
+
+        error = None
+        user = myDB.execute(
+            'SELECT * FROM user WHERE username = ?', (username,)
+        ).fetchone()
+
+        if user is None:
+            error = 'Incorrect username.'
+        elif not check_password_hash(user['password'], password):
+            error = 'Incorrect password.'
+
+        if error is None:
+            session.clear()
+            session['user_id'] = user['username']
+            return redirect(url_for('list_vms'))
+
+        flash(error, "error")
+
+    return render_template('login.html')
+
+
+@app.route("/logout")
+def logout():
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
+    session.pop("user_id")
+    return redirect(url_for('login'))
+
+
+@app.route('/register', methods=('GET', 'POST'))
+def register():
+    if request.method == 'POST':
+        username = request.form['username']
+        password = request.form['password']
+        myDB = db.get_db()
+
+        error = None
+
+        if not username:
+            error = 'Username is required.'
+        elif not password:
+            error = 'Password is required.'
+        elif myDB.execute(
+            'SELECT username FROM user WHERE username = ?', (username,)
+        ).fetchone() is not None:
+            error = 'User {} is already registered.'.format(username)
+
+        if error is None:
+            myDB.execute(
+                'INSERT INTO user (username, password) VALUES (?, ?)',
+                (username, generate_password_hash(password))
+            )
+            myDB.commit()
+            return redirect(url_for('hello_world'))
+
+        flash(error, "error")
+
+    return render_template('register.html')
 
 
 @app.route("/list_vms")
 def list_vms():
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
+    libvirt_instance.update_dom_dict()
+    myDB = db.get_db()
+    vms = db.get_vms_by_user(myDB, session["user_id"])
+    vms_dict = {}
+    for vm in vms:
+        vms_dict[vm] = libvirt_instance.domains[vm]
+    return render_template("list_vms.html", vm_dict=vms_dict)
+
+
+@app.route("/list_all_vms")
+def list_all_vms():
+    if ("user_id" not in session) or (session["user_id"] != "admin"):
+        return redirect(url_for('list_vms'))
+
     libvirt_instance.update_dom_dict()
     return render_template("list_vms.html", vm_dict=libvirt_instance.domains)
 
 
 @app.route("/start/<vm_name>")
 def start_vm(vm_name):
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
     try:
         libvirt_instance.start_domain(dom_name=vm_name)
         flash("VM started successfully", category="success")
@@ -33,6 +139,9 @@ def start_vm(vm_name):
 
 @app.route("/stop/<vm_name>")
 def stop_vm(vm_name):
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
     try:
         libvirt_instance.stop_domain(dom_name=vm_name)
         flash("VM stopped successfully", category="success")
@@ -44,6 +153,9 @@ def stop_vm(vm_name):
 
 @app.route("/delete/<vm_name>")
 def delete_vm(vm_name):
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+
     try:
         libvirt_instance.delete_domain(dom_name=vm_name)
         flash("VM deleted successfully", category="success")
@@ -55,13 +167,19 @@ def delete_vm(vm_name):
 
 @app.route("/create", methods=['GET', 'POST'])
 def create_vm():
+    if "user_id" not in session:
+        return redirect(url_for('login'))
+    myDB = db.get_db()
     if request.method == "GET":
         return render_template("create_vm.html")
     elif request.method == "POST":
         try:
-            libvirt_instance.create_domain(dom_name=request.form.get("vm_name"),
-                                           num_cpu=request.form.get("cpu_num"),
-                                           mem=request.form.get("mem_size"))
+            dom = libvirt_instance.create_domain(dom_name=request.form.get("vm_name"),
+                                                 num_cpu=request.form.get("cpu_num"),
+                                                 mem=request.form.get("mem_size"))
+
+            db.reg_domain(myDB, dom.name(), session["user_id"])
+
             flash("VM created successfully", category="success")
         except Exception as e:
             flash(message=str(e), category="warning")
